AWK solution for Day 9.
I messed this one up because I made a call to
x_adjust_tail, updating the value of the array, and then called
y_adjust_tail on the value that was updated instead of the original.
Needless to say, I was very disappointed by this mistake but am glad
that the issue was resolved by using global variables instead of
recalculating x_dist and y_dist a second time when it was time to move
the y-coordinate.